+++ quic/s2n-quic-core/tests/crypto/fuzz_target.rs
@@ -129,11 +129,11 @@ impl Key for FuzzCrypto {
         0
     }
 
-    fn aead_confidentiality_limit(&self) -> usize {
+    fn aead_confidentiality_limit(&self) -> u64 {
         0
     }
 
-    fn aead_integrity_limit(&self) -> usize {
+    fn aead_integrity_limit(&self) -> u64 {
         0
     }
 }
