+++ quic/s2n-quic-ring/src/negotiated.rs
@@ -66,11 +66,11 @@ impl Key for RingNegotiatedCrypto {
         self.sealer.tag_len()
     }
 
-    fn aead_confidentiality_limit(&self) -> usize {
+    fn aead_confidentiality_limit(&self) -> u64 {
         self.sealer.aead_confidentiality_limit()
     }
 
-    fn aead_integrity_limit(&self) -> usize {
+    fn aead_integrity_limit(&self) -> u64 {
         self.sealer.aead_integrity_limit()
     }
 }
@@ -149,11 +149,11 @@ macro_rules! negotiated_crypto {
                 self.0.tag_len()
             }
 
-            fn aead_confidentiality_limit(&self) -> usize {
+            fn aead_confidentiality_limit(&self) -> u64 {
                 self.0.aead_confidentiality_limit()
             }
 
-            fn aead_integrity_limit(&self) -> usize {
+            fn aead_integrity_limit(&self) -> u64 {
                 self.0.aead_integrity_limit()
             }
         }
